updated component structure for easier props flow
